gfx,tools: Compress joint influence data further.
Since joint indices are local anyway, we don't need too many bits
for these, and 11 bits of precision should be enough for weights
as well considering that GLTF allows to go down to 8.

Cuts the amount of memory needed for this in half, which can be
significant if there are highly detailed joint animations.